+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: unifir
 Title: A Unifying API for Calling 'Unity' from R
-Version: 0.1.0.9001
+Version: 0.1.0
 Authors@R: 
     person("Michael", "Mahoney", , "mike.mahoney.218@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2402-304X"))
