### 2022-03-24 #521 @mikemahoney218
While it's not relevant to this issue, I do wonder if `has [continuous integration](https://ropensci.github.io/dev_guide/ci.html), including reporting of test coverage using services such as Travis CI, Coveralls and/or CodeCov.` should be updated to remove Travis now that rOpenSci recommends moving away from Travis?
https://devguide.ropensci.org/ci.html?q=travis#travis-ci-linux-and-mac-osx
