+++ man/create_unity_project.Rd
@@ -12,8 +12,8 @@ create_unity_project(path, quit = TRUE, unity = NULL)
 \item{quit}{Logical: quit Unity after creating the project?}
 
 \item{unity}{The path to the Unity executable on your system (importantly,
-_not_ the UnityHub executable). If `NULL`, checks to see if the environment
-variable or option `unifir_unity_path` is set; if so, uses that path
+\emph{not} the UnityHub executable). If \code{NULL}, checks to see if the environment
+variable or option \code{unifir_unity_path} is set; if so, uses that path
 (preferring the environment variable over the option if the two disagree).}
 }
 \value{
