+++ man/waiver.Rd
@@ -9,8 +9,8 @@ waiver()
 \description{
 This function is borrowed from ggplot2. It creates a "flag" object indicating
 that a value has been intentionally left blank (because it will be filled in
-by something else). Often, a function argument being missing or `NULL` will
-result in an error, while passing `waiver()` will cause the function to look
+by something else). Often, a function argument being missing or \code{NULL} will
+result in an error, while passing \code{waiver()} will cause the function to look
 elsewhere in the script for an acceptable value.
 }
 \examples{
