+++ man/set_active_scene.Rd
@@ -7,8 +7,8 @@
 set_active_scene(script, scene_name = NULL, method_name = NULL, exec = FALSE)
 }
 \arguments{
-\item{script}{A `unifir_script` object, created by [make_script] or returned
-by an `add_prop_*` function.}
+\item{script}{A \code{unifir_script} object, created by \link{make_script} or returned
+by an \verb{add_prop_*} function.}
 
 \item{scene_name}{The name of the scene to set as the active scene.}
 
