### 2022-04-16 #521 @vinhtantran
## Package Review

-   **Briefly describe any working relationship you have (had) with the package authors.** None
-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

I am not a Unity user or game developer, so I have to rely to the *“Why do this?”* section in the README to understand why the package is needed. The author pointed out two motivations: the use of closed-source tooling in current research and the potential of bias for personal preference. The first motivation is intriguing in that the package has the potential to encourage researchers to publish open-source and reproducible source codes for their studies. However, I hope the author can elaborate on the second point in terms of how the package would assist with the aforementioned issue.

The source code was well-written and packaged. I have nothing to say about how the code was written and structured. It’s not surprising given that this is not the author’s first submission to rOpenSci.

The vignettes are simple to follow and put into action. By following the 102 vignette, I was able to create a toy working example in Unity. I understand that the CRAN check and GitHub actions limit your ability to showcase more sophisticated examples with your assets, but I’d love to see one that highlights all package’s features, perhaps from a blogpost? It’s only a suggestion.

Another point I’d like to make is about the package’s ongoing development. I cloned the package and wrote a bunch of comments on some inconsistencies in your function documentation a few weeks ago, only to discover that the majority of them have been fixed in your most recent commits, which were made a few days ago. Although the majority of them are minor changes, I would recommend that you include a note in the Review report to let us reviewers know which version we should work on.

Finally, congratulations on a job well done. The review process taught me a lot, including new information about the R6 object in R.

I’ve included some detailed comments below. Some of them may be obsolete in your most recent commit because I did not have time to revise them all.

#### Detailed Comments

-   Function documentation:
    -   Include a link to the referred function in the Descriptions, such as the reference to `instantiate_prefab()` in `add_default_player()`.
    -   Make use of fixed width font in Markdown by putting in \` \` pairs when referring to function and argument names.
    -   The object names displayed are not consistent. For example, in `add_texture()`, it states “A ‘unifir\_script’ object”, then in `add_prop()`, it states “A \[unifir\_prop\] object”.
    -   All sentences in “Description” should end with a period.
    -   `find_unity()` Description is mixed with Details. They are not continuous. Either move everything to the Description or the Details sections.
-   Vignettes:
    -   I was able to follow the vignettes step-by-step and helped me understand the package and its purpose because I have not known Unity before.
    -   The `library(unifir)` in the setup chunk at top of the pages should be hidden (`echo = FALSE`). Readers should see it right before the first call to the package’s function, similar to what you have in “Unifir 101”, instead of out of nowhere at the top of the pages.
    -   The notice of usage of `waiver()` and others to ensure rendering on CRAN could be put in the comment on top of the function call instead of the text below the code. Also, make sure to note all those “blank code”. You miss one `waiver()` notice at Action section of 102.
-   Function Examples:
    -   `get_asset()` calls `add_players()` which is no longer available in the latest commit.
-   Packaging Guidelines:
    -   I’d suggest removing ‘from R’ from the Title to avoid error when publishing to CRAN.
