### 2022-04-25 mikemahoney218/unifir#4 @wjones127
Issue finding Unity on MacOS and Windows
Had issues finding Unity on MacOS. This is what I saw in my interactive session:

```r
library(unifir)

options(unifir_unity_path="/Applications/Unity/Hub/Editor/2021.3.0f1/Unity.app/Contents/MacOS/Unity")
find_unity() # success on first call
# [1] "\"/Applications/Unity/Hub/Editor/2021.3.0f1/Unity.app/Contents/MacOS/Unity\""
find_unity() # failure on second call
# Error in find_unity() : Couldn't find Unity executable at provided path.
# Please make sure the path provided to 'unity' is correct.
Sys.getenv("unifir_unity_path")
# [1] "\"/Applications/Unity/Hub/Editor/2021.3.0f1/Unity.app/Contents/MacOS/Unity\""
```

It works if I remove the logic that adds quotation marks to the path and add a beginning slash on this line:

https://github.com/mikemahoney218/unifir/blob/2e5f90cfc09b4456b3751f932f19da49e0dbebc7/R/find_unity.R#L86

On Windows, the `windows_location()` works fine, but `file.exists(unity)` always returns false with the extra `"` added to the path. I think this means even if the user profiles a different unity path, it will always select the default system one.

Relatedly, if I run the test locally on MacOS, I get this failure:

```r
Failure (test-make_script.R:64:3): make_script always produces the same script
example_script$unity (`actual`) not identical to saved_script$unity (`expected`).

`actual`:   "junk_string"
`expected`: "\"junk_string\""
```
