### 2022-04-25 #521 @wjones127
I'm about half-way through my review, but in the meantime I've posted an issue in the repo:

 * https://github.com/mikemahoney218/unifir/issues/4
