### 2022-04-26 ropensci/unifir#4 @mikemahoney218
Gah! Thank you, will look into this today. Any chance you could post your `sessionInfo()`?
