### 2022-04-26 ropensci/unifir#4 @wjones127
Sure!

```
R version 4.1.3 (2022-03-10)
Platform: aarch64-apple-darwin20 (64-bit)
Running under: macOS Monterey 12.3.1

Matrix products: default
LAPACK: /Library/Frameworks/R.framework/Versions/4.1-arm64/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
[1] unifir_0.1.0

[1] compiler_4.1.3        R6_2.5.1              tools_4.1.3           glue_1.6.2
[5] proceduralnames_0.2.1
```
