+++ tests/testthat/test-not_on_ci.R
@@ -1,4 +1,4 @@
-test_that("multiplication works", {
+test_that("skipped test", {
   skip_on_ci()
   skip_on_cran()
   Sys.setenv("unifir_debugmode" = "")
