### 2022-04-26 ropensci/unifir#4 @mikemahoney218
Thanks! I made some changes in 128714fc524fd9244d2a9000d2429af3f81967f7 to hopefully address this (on the branch `response_to_reviewers`), though I haven't gotten to test them yet.
