### 2022-04-26 #521 @wjones127
Added two other issues:

 * https://github.com/mikemahoney218/unifir/issues/6
 * https://github.com/mikemahoney218/unifir/issues/7
