### 2022-04-26 bfbb306
Handle non-local terrain (fixes #6)
