### 2022-04-26 ropensci/unifir#6 @mikemahoney218
Thanks! Fixed in bfbb306 -- I had been testing with local files and mistakenly assumed that I could swap the local path for a tempfile (for CRAN) with no problems.

Unfortunately, that's actually the entire error message provided by Unity when called on the command line. There might be a space for a vignette on how to debug these scripts (which is a bit more about the Unity engine and C# than anything to do with R) but Unity's CLI is notoriously unfriendly, so I don't believe there's any way to surface these issues within R itself.
