### 2022-04-26 mikemahoney218/unifir#8 @wjones127
associated_coordinates() example errors due to missing CRS
Here's the reprex I ran:

```r
library(unifir)
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)
simulated_data <- sf::st_as_sf(simulated_data, coords = c("lng", "lat"))
output_files <- terrainr::get_tiles(simulated_data)
#> Warning in get_tiles.sf(simulated_data): Assuming geographic CRS. Set
#> 'projected' to TRUE if projected.
#> Warning in handle_bboxSR(data, projected): Assuming CRS of EPSG 4326 (set bboxSR
#> explicity to override)
temptiff <- tempfile(fileext = ".tif")
terrainr::merge_rasters(output_files["elevation"][[1]], temptiff)
associate_coordinates(simulated_data, temptiff)
#> Error in st_transform.sfc(st_geometry(x), crs, ...): cannot transform sfc object with missing crs
```
