+++ tests/testthat/test-create_terrain.R
@@ -108,7 +108,7 @@ test_that("create_terrain actions as expected", {
 
   writeLines(gsub(" |\\n|\\r", "", outcome$props[[4]]), actual)
   writeLines(
-    gsub(" |\\n|\\r", "", "static void TerrainTest()\n{\n    string heightmapPath = \"heightmap_test\";\n    float x_pos = 0F;\n    float z_pos = 0F;\n    float width = 4097F;\n    float height = 4097F;\n    float length = 4097F;\n    int heightmapResolution = 4097;\n    string texturePath = \"texture_test\";\n\n    TerrainData terrainData = new TerrainData();\n    terrainData.size = new Vector3(width / 128, height, length / 128);\n    terrainData.heightmapResolution = heightmapResolution;\n\n    GameObject terrain = (GameObject)Terrain.CreateTerrainGameObject(terrainData);\n    terrain.transform.position = new Vector3(x_pos, 0, z_pos);\n    float [,] heights = ReadRawAutoAdd(heightmapPath, heightmapResolution);\n    terrainData.SetHeights(0, 0, heights);\n    if(texturePath != string.Empty){\n        AddTextureAutoAdd(texturePath, terrainData, width, length);\n    }\n    AssetDatabase.CreateAsset(terrainData, \"Assets/\" + heightmapPath + \".asset\");\n}\n"), # nolint
+    "staticvoidTerrainTest(){stringheightmapPath=\"heightmap_test\";stringbasePath=\"heightmap_test\";floatx_pos=0F;floatz_pos=0F;floatwidth=4097F;floatheight=4097F;floatlength=4097F;intheightmapResolution=4097;stringtexturePath=\"texture_test\";TerrainDataterrainData=newTerrainData();terrainData.size=newVector3(width/128,height,length/128);terrainData.heightmapResolution=heightmapResolution;GameObjectterrain=(GameObject)Terrain.CreateTerrainGameObject(terrainData);terrain.transform.position=newVector3(x_pos,0,z_pos);float[,]heights=ReadRawAutoAdd(heightmapPath,heightmapResolution);terrainData.SetHeights(0,0,heights);if(texturePath!=string.Empty){AddTextureAutoAdd(texturePath,terrainData,width,length);}AssetDatabase.CreateAsset(terrainData,\"Assets/\"+basePath+\".asset\");}", # nolint
     expected
   )
 
