### 2022-04-26 ropensci/unifir#4 @wjones127
Great. I can confirm that branch fixes the issue for me. The unit tests pass and the Unity path is consistently found on MacOS.
