### 2022-04-26 1d5b1fe
Fix examples (fixes #8)
