### 2022-04-26 ropensci/unifir#8 @mikemahoney218
Addressed this in 1d5b1fe -- in addition to explicitly setting the CRS in this example so that it's now correct, I also added CI to run the examples so that failing examples wrapped in `dontrun` will still get caught. Thanks for flagging!
