### 2022-04-26 ropensci/unifir#6 @wjones127
Can confirm that this example now works in that fix branch.

> Unity's CLI is notoriously unfriendly, so I don't believe there's any way to surface these issues within R itself.

That's a shame 😢
