### 2022-04-26 #521 @wjones127
---

Overall this package is very good. The code style and level of testing are excellent.

I worked through the vignettes successfully, and then tried to create a simple scene of my own. I've reported the major issues I encountered in that process, which I think ought to be addressed (most already have 💯 ):

* https://github.com/mikemahoney218/unifir/issues/4
* https://github.com/mikemahoney218/unifir/issues/6
* https://github.com/mikemahoney218/unifir/issues/7
* https://github.com/mikemahoney218/unifir/issues/8

In addition, I noticed a few smaller issues while reviewing the package:

* [ ] `find_unity()` docs are a little funny: the enumerated list is pushed down to details, so it's a little confusing.
* [ ] The readme doesn't have linked to the vignettes, though that is suggested in the rOpenSci guidelines.
* [ ] By default, `add_default_tree()` adds trees that are 90 degrees from vertical.
