+++ R/action.R
@@ -1,6 +1,7 @@
 #' Build and execute a `unifir_script`
 #'
-#' @param script The `unifir_script` object to build and execute.
+#' @param script The `unifir_script` object (as generated by [make_script])
+#' to build and execute.
 #' @param write Boolean: Write the generated script to a file?
 #' @param exec Boolean: Execute the script inside of the Unity project? Note
 #' that if `write = FALSE`, `exec` cannot be `TRUE`.
@@ -23,6 +24,10 @@
 #'   action(script)
 #' }
 #' }
+#'
+#' @return If `exec = FALSE`, the original `unifir_script` object passed to
+#' `script`. If `exec = TRUE`, the same `unifir_script` object with its
+#' props replaced by the C# they generate.
 #'
 #' @export
 action <- function(script, write = TRUE, exec = TRUE, quit = TRUE) {
