+++ R/scene_management.R
@@ -81,9 +81,7 @@ new_scene <- function(script,
 #' # First, create a script object.
 #' # CRAN doesn't have Unity installed, so pass
 #' # a waiver object to skip the Unity-lookup stage:
-#' script <- make_script("example_script",
-#'   unity = waiver()
-#' )
+#' script <- make_script("example_script", unity = waiver())
 #'
 #' # Now add props:
 #' script <- load_scene(script, scene_name = "some_scene")
