+++ R/associate_coordinates.R
@@ -5,12 +5,11 @@
 #' can only import terrain as square tiles of side 2^x + 1, for x between 5 and
 #' 12. As a result, importing objects into a Unity scene so that they align
 #' with terrain surfaces is trickier than you'd expect.
-#'
 #' This function "associates" the XY coordinates from some `sf` object, likely a
 #' point data set, with some raster object.
 #'
 #' @param object The `sf` object to take coordinates from. The object will be
-#' reprojected to align with `raster`.
+#' reprojected (via [sf::st_transform]) to align with `raster`.
 #' @param raster A raster or file path to a raster to associate coordinates
 #' with. Note that different rasters will produce different coordinate outputs;
 #' you should run this function with the same raster you plan on bringing into
@@ -22,7 +21,7 @@
 #' @return A data.frame with two columns, X and Y, representing the re-aligned
 #' coordinates. If `object` is point data (or anything object that
 #' [sf::st_coordinates] returns a single row for each row of), these rows will
-#' be in the same order as `object` (and so can be appended via `cbind`).
+#' be in the same order as `object` (and so can be appended via [cbind]).
 #'
 #' @examples
 #' \dontrun{
