### 2022-04-26 #521 @mikemahoney218
Will do! I've got a feature branch with changes in response to these reviews open at https://github.com/mikemahoney218/unifir/pull/5 , and I expect to have addressed all comments in the near future -- I'll update here once I've done so.

Thanks everyone!
