### 2022-04-28 1e2b056
Rotate trees correctly if not specified
