+++ vignettes/unifir-user-guide.Rmd
@@ -124,6 +124,10 @@ if we look at `script$props` instead, we'd see that we now have three
 `unifir_prop` objects in our script. This list is how unifir keeps track of what
 exactly it needs to do to make your script into a scene; it will run through 
 your props in the order you added them to the script.
+`add_default_player()` is just one function that adds props to our script; 
+other functions in unifir will add props to add lights, 3D models, terrain,
+and more to your script. You can see the prop functions listed under "See Also" 
+in any prop's help page (so for instance, after running `?add_default_player`).
 
 One thing to highlight is that we need to explicitly save our scenes at the end
 of any script. If you don't save your scene, then after your script executes
@@ -133,6 +137,18 @@ to add a `save_scene()` call to the end of your script:
 ```{r eval = FALSE}
 script <- save_scene(script)
 ```
+
+This will save our scene with all of its changes, but Unity won't automatically 
+default to _loading_ this scene when we open our project. If we want our scene
+to load as soon as we open the project, we can add another prop to our script
+via the function `set_active_scene()`:
+
+```{r eval=FALSE}
+script <- set_active_scene(script)
+```
+
+If you forget to do this, you can open scenes from inside Unity
+(File -> Open Scene); unifir always saves scenes to the "Scenes" folder.
 
 Now that we've added props to our script, it's time to actually make our
 scene a reality. You can play out your script using the function `action()`,
@@ -148,6 +164,179 @@ around in! Note that your scene won't be open by default (unless you've used
 the `set_active_scene()` function) -- to actually _see_ all the changes you've
 made, make sure to load the scene located in the `Scenes` folder inside your
 project.
+
+## A full example
+
+So now that we know the basics of how a unifir script is built, we should walk
+through how to actually build one to make a scene! We'll walk through adding
+terrain, trees, and a character to a scene and then see what it looks like when
+rendered. This example comes from [Will Jones](https://github.com/wjones127), 
+who helped 
+[improve this vignette](https://github.com/mikemahoney218/unifir/issues/7) 
+during 
+[unifir's review at rOpenSci](https://github.com/ropensci/software-review/issues/521).
+
+First things first, we'll need a script! Just like before, we'll make our script
+using the `make_script()` function:
+
+```{r eval = FALSE}
+tree_script <- make_script(
+  project = file.path(tempdir(), "unifir", "random_trees")
+  )
+```
+
+Up next, we need data on the terrain we want to add! We can generate
+a random terrain surface using the terra package, so we'll load that now:
+
+```{r eval = FALSE}
+library(terra)
+```
+
+And we'll need to convert that terrain into a format Unity can read, so we'll
+load in the terrainr package as well:
+
+```{r eval = FALSE}
+library(terrainr)
+```
+
+Importing terrain into Unity is a little bit wonky. First off, Unity can't
+process standard raster data, but rather needs a specialized format that
+most libraries don't work with; for that reason, we're going to need to use
+terrainr's `transform_elevation()` function to transform our elevation raster 
+into a proper format.
+
+Additionally, Unity doesn't have a great way for wrangling objects of different
+unit systems -- it assumes that a distance of "1" in a 3D model is also a
+distance of "1" in the terrain, no matter what units that "1" is in.
+It's common, but not at all universal, for a distance of "1" to represent a 
+distance of 1 meter, but depending where you're getting assets from you might
+need to rescale your rasters and models to make everything line up 
+appropriately.
+
+Lastly, Unity expects terrain tiles to be (2^x) + 1 unit squares, where x is
[…]
