### 2022-04-28 ropensci/unifir#7 @mikemahoney218
Two comments on things brought up here:

1. The supported way to get the player controller on the ground would be to have a point (in the raster's CRS) where you want the player to be, use terra::extract to get the elevation at that point, and then associate_coordinates to place the player in the scene; I'm attempting to keep unifir as not-spatial as possible so it's a "lighter" installation (hence terra and sf in Suggests) and easier to incorporate into other packages
2. Fixed the tree rotation thing, that was silly on my end (1e2b05630afa09df3a0527c397496c7770cd0d8e) (and is related to the coordinate system thing -- most 3D modeling programs use a different coordinate system, so their models aren't properly rotated for Unity https://www.techarthub.com/a-guide-to-unitys-coordinate-system-with-practical-examples/ )

Otherwise, I've added a lot to the user vignette in bb1c2185e6b0252322940bdeaba570e23c4b8988 .
