### 2022-04-28 3d409d0
Remove unused NEWS sections
