### 2022-04-28 #521 @mikemahoney218
Hi all! I believe I've addressed all reviewer comments in a branch on the unifir repo, https://github.com/mikemahoney218/unifir/tree/response_to_reviewers (diff at https://github.com/mikemahoney218/unifir/pull/5/files , NEWS at https://github.com/mikemahoney218/unifir/blob/response_to_reviewers/NEWS.md ). Thank you so much for your comments, the package is looking a lot better as a result!

# Response to @vinhtantran

Thanks so much for your review! I'd like to include you in the DESCRIPTION as a reviewer -- would you be able to tell me what I should list for "given" and "family" names? I don't want to infer incorrectly!

## Function documentation

* [x] Include a link to the referred function in the Descriptions, such as the reference to instantiate_prefab() in add_default_player()
  * Believe this is now fixed, thanks!
* [x] Make use of fixed width font in Markdown by putting in ` ` pairs when referring to function and argument names. (AND) [x] The object names displayed are not consistent. For example, in add_texture(), it states “A ‘unifir_script’ object”, then in add_prop(), it states “A [unifir_prop] object”.
  * I've done another pass through all the function documentation to try and address this -- classes are now demarcated by ``, functions by [ (which should generate links), and I now attempt to always refer to both the relevant class and constructor function (so the given example now reads "A `unifir_script` object created by [make_script]" and "a `unifir_prop` object created by [unifir_prop]".
* [x] All sentences in “Description” should end with a period.
  * Fixed! Thanks for flagging.
* [x] find_unity() Description is mixed with Details. They are not continuous. Either move everything to the Description or the Details sections.
  * Whoops! Messed that up. Thanks for flagging.

## Vignettes

* [x] The library(unifir) in the setup chunk at top of the pages should be hidden (echo = FALSE). Readers should see it right before the first call to the package’s function, similar to what you have in “Unifir 101”, instead of out of nowhere at the top of the pages.
  * Fixed!
* [x] The notice of usage of waiver() and others to ensure rendering on CRAN could be put in the comment on top of the function call instead of the text below the code. Also, make sure to note all those “blank code”. You miss one waiver() notice at Action section of 102.
  * Done!

## Packaging guidelines

* [x] I’d suggest removing ‘from R’ from the Title to avoid error when publishing to CRAN.
  * Done! Thanks for flagging.

---

# Response to @wjones127

Thank you so much for your review! Below I'm only addressing the comments on this issue; I've dealt with the larger issues on their own tickets in the unifir repo. I'd like to include you in the DESCRIPTION as a reviewer, as well -- should I list you as Will Jones or something else?

* [x] `find_unity()` docs are a little funny: the enumerated list is pushed down to details, so it's a little confusing.
  * Fixed! Thanks. Guess I misunderstood how roxygen2 generates these sections.
* [x] The readme doesn't have linked to the vignettes, though that is suggested in the rOpenSci guidelines.
  * Fixed!
* [x] By default, `add_default_tree()` adds trees that are 90 degrees from vertical.
  * Fixed!
