+++ paper.bib
@@ -1,0 +1,79 @@
+@inproceedings{Slater1993,
+  author={Slater, M. and Usoh, M.},
+  booktitle={Proceedings of IEEE Virtual Reality Annual International Symposium}, 
+  title={Presence in immersive virtual environments}, 
+  year={1993},
+  volume={},
+  number={},
+  pages={90-96},
+  doi={10.1109/VRAIS.1993.380793}}
+
+@article{Schone2019,
+  author          = {Schöne, B. and Wessels, M. and Gruber, T},
+  journal         = {Current Psychology},
+  title           = {Experiences in Virtual Reality: a Window to Autobiographical Memory},
+  volume          = {38},
+  year            = {2019},
+  pages           = {715-719},
+  doi             = {10.1007/s12144-017-9648-y},
+}
+
+@article{Huang2021,
+author = {Jiawei Huang and Melissa S. Lucash and Robert M. Scheller and Alexander Klippel},
+title = {Walking through the forests of the future: using data-driven virtual reality to visualize forests under climate change},
+journal = {International Journal of Geographical Information Science},
+volume = {35},
+number = {6},
+pages = {1155-1178},
+year  = {2021},
+publisher = {Taylor & Francis},
+doi = {10.1080/13658816.2020.1830997},
+}
+
+@article{Keil2020,
+  author          = {Keil, Julian and Edler, Dennis and Schmitt, Thomas and Dickmann, Francis},
+  title           = {Creating Immersive Virtual Environments Based on Open Geospatial Data and Game Engines},
+  journal         = {KN Journal of Cartography and Geographic Information},
+  volume          = {71},
+  year            = {2021},
+  pages           = {53-65},
+}
+
+  @Article{terrainr,
+    year = {2022},
+    publisher = {The Open Journal},
+    volume = {7},
+    number = {69},
+    pages = {4060},
+    author = {Michael J. Mahoney and Colin M. Beier and Aidan C. Ackerman},
+    title = {{terrainr}: An R package for creating immersive virtual environments},
+    journal = {Journal of Open Source Software},
+    doi = {10.21105/joss.04060},
+    url = {https://doi.org/10.21105/joss.04060},
+  }
+
+@article{Swetnam2019,
+title = {Can computer game landscapes target new audiences for landscape quality assessment?},
+journal = {Applied Geography},
+volume = {113},
+pages = {102102},
+year = {2019},
+issn = {0143-6228},
+doi = {https://doi.org/10.1016/j.apgeog.2019.102102},
+url = {https://www.sciencedirect.com/science/article/pii/S0143622818312207},
+author = {R.D. Swetnam and J. Korenko},
+}
+
+@article{Fiore2009,
+title = {Virtual experiments and environmental policy},
+journal = {Journal of Environmental Economics and Management},
+volume = {57},
+number = {1},
+pages = {65-86},
+year = {2009},
+note = {Frontiers of Environmental and Resource Economics},
+issn = {0095-0696},
+doi = {https://doi.org/10.1016/j.jeem.2008.08.002},
+url = {https://www.sciencedirect.com/science/article/pii/S0095069608000983},
+author = {Stephen M. Fiore and Glenn W. Harrison and Charles E. Hughes and E. Elisabet Rutström},
+}
