+++ paper.Rmd
@@ -1,0 +1,151 @@
+---
+title: 'unifir: A Unifying API for Working with Unity in R'
+tags:
+  - R
+  - virtual reality
+  - GIS
+  - Unity
+authors:
+  - name: Michael J. Mahoney
+    orcid:  0000-0003-2402-304X 
+    affiliation: 1
+  - name: Colin M. Beier
+    orcid:  0000-0003-2692-7296 
+    affiliation: 2
+  - name: Aidan C. Ackerman
+    orcid:   0000-0001-7106-0957 
+    affiliation: 3
+affiliations:
+ - name: Graduate Program in Environmental Science, State University of New York College of Environmental Science and Forestry, Syracuse, New York, USA
+   index: 1
+ - name: Department of Sustainable Resources Management, State University of New York College of Environmental Science and Forestry, Syracuse, New York, USA
+   index: 2
+ - name: Department of Landscape Architecture, State University of New York College of Environmental Science and Forestry, Syracuse, New York, USA
+   index: 3
+citation_author: Mahoney et. al.
+date: "`r Sys.Date()`"
+year: 2022
+bibliography: paper.bib
+output: rticles::joss_article
+csl: apa.csl
+journal: JOSS
+---
+```{r setup, echo = FALSE, message=FALSE}
+knitr::opts_chunk$set(
+  echo = FALSE,
+  message = FALSE
+)
+library(dplyr)
+library(ggplot2)
+library(kableExtra)
+```
+
+# Summary
+
+unifir is an R package representing a first step towards an open-source, 
+fully reproducible approach 
+for constructing immersive virtual environments (IVEs). 
+Users are able to construct IVEs from real or simulated spatial data 
+by writing simple, idiomatic R code, allowing the production of IVEs without
+needing users to be highly familiar with the complicated video game engines
+used to power these environments. 
+unifir also provides a simple set of classes which are easily extended, 
+enabling future efforts to provide users with a consistent API for IVE design
+while relying upon unifir as a low-level interface to the Unity game engine and 
+extending its capabilities to new data formats or engine capabilities.
+
+# Statement of Need
+
+Immersive virtual environments (IVEs) are an exciting area of research 
+with potential experimental applications in fields varying from
+science communication [@Huang2021],
+landscape planning [@Swetnam2019],
+environmental economics [@Fiore2009]
+and beyond.
+By creating realistically-rendered immersive representations, 
+researchers are able to induce a sense of presence (or "being there") in 
+their audiences [@Slater1993], 
+which may serve to make experiments in IVEs
+more similar to "real-life" experiences
+than can normally be obtained through controlled experiments 
+in laboratory settings [@Schone2019].
+However, despite their utility, the difficulty of producing IVEs has limited
+their use, as the creation of IVEs relies upon a number of software tools which
+are not commonly used in standard scientific practice. 
+These tools are often managed through graphical user interfaces, making it
+difficult to capture the production process in a way that would allow 
+others to reproduce IVEs or replicate IVE-driven experiments. 
+Finally, it is also difficult for researchers to incorporate spatial data 
+into IVEs using these tools 
+in order to produce virtual representations of real-world places
+[@Keil2020].
+
+The core gap underpinning all of these challenges is a lack of accessible 
+tooling providing a standardized, programmatic method for creating IVEs from
+real or simulated data. Some first efforts have been made in this direction; 
+for instance, Keil et al. [-@Keil2020] surveyed methods for incorporating 
+spatial data representing both terrain and buildings into IVEs,
+while Mahoney et al. [-@terrainr] introduced an open-source toolkit for 
+programmatically incorporating terrain data into IVEs. 
+More work is still needed however to both expand the amount of IVE production
[…]
