### 2022-04-29 #521 @wjones127
@mikemahoney218 I can confirm that all my comments have been addressed. I left a few optional follow-up comments in my final review of the changes here: https://github.com/mikemahoney218/unifir/pull/5#pullrequestreview-957918764

Great work!

> I'd like to include you in the DESCRIPTION as a reviewer, as well -- should I list you as Will Jones or something else?

Yes, "Will Jones" is good to list me as. Email is willjones127 at gmail, if you need that.
