### 2022-04-29 76d8912
Apply suggestions from code review
Co-authored-by: Will Jones <willjones127@gmail.com>
