### 2022-04-29 a5f5bad
Add CRAN comments and re-render codemeta
