+++ cran-comments.md
@@ -1,0 +1,15 @@
+## Test environments
+* local R installation, R 4.2.0
+* ubuntu 20.04 (on GitHub Actions), R devel, release, oldrelease
+* MacOS X 11.6.3 (on GitHub Actions), R release
+* Windows Server 2019 (on GitHub Actions), R release
+* win-builder (devel, release, oldrelease)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
+
+## Downstream dependencies
+There are no downstream dependencies on CRAN at the time of submission.
