### 2022-04-29 8e543dd
Small tweaks for release version (#9)
* Spell check package

* Add CRAN comments and re-render codemeta

* Wrap download
