### 2022-05-01 #521 @melvidoni
Hello @mikemahoney218. You'll have to wait so I can ask about the bot. I believe it should have created the group.
Give me a couple days, please.
