### 2022-05-02 #521 @mikemahoney218
Awesome, thanks @maelle ! In case it helps debugging, I manually requested access to the team after the bot didn't add me, I didn't get added without approval.

I believe that means the repo is transferred and we're all good to go! Thank you so much everyone!
